feat(bcgov/storbook): update grid
change to mdx format

#294
